fixed window not un-hiding widgets in fullscreen
A separate resizeEvent has been added for dockControls which exclusively
handles hiding and showing widgets based on the size of the controls.
This corrects issues where the dockControls would not un-hide widgets
when entering fullscreen mode, as well as any other edge-cases where
dockControls might have otherwise missed a resizeEvent.