Added theme-aware spacing, further tightened UI
Currently only for frameProgress. The maximum size of frameProgress is
adjusted based on the size of the handle and groove of the progress bar
in the currently selected theme. UI has been slightly compacted further,
and the statusbar in particular now has a maximum height, taking up far
less room. Trim buttons are now aligned with the rest of the UI.