Fixed QWidgetPassthrough and promoted lineOutput
Fixes bug with QWidgetPassthrough that caused all key pressed to be
ignored due to subclasses not accessing the correct keyPressEvent with
super() by adding a class variable called "base" to represent the class's
true inheritance. Not sure if it's a bandaid fix or the actual solution.
Also, added QLineEditPassthrough variant for the output text box so
users can use Esc to clear focus, as well as a setIgnoreAll() method to
immediately toggle all categories (but not ignoredKeys).