Improved cross-platform ffmpeg/PATH detection
qthelpers.file_in_PATH now works cross-platform and returns the full
path to a file, rather than just the directory. constants.verify_ffmpeg
is more robust, actually updates the path when needed, has better
cross-platform support, and logs a unique message for if ffmpeg was
not detected, but no QMessageBox warning was shown.