Improved/refactored thumbnail-generation
Uses a much, much faster method for generating thumbnails
that is not dependent on the length of the video. Thumbnail
generation has been refactored in general to improve readablility
and robustness. The thumbnail folder is recreated when needed,
and filenames now only use the basename of their video, fixing
bugs with certain paths.