Refactored updater, added help message
This adds short arguments to the updater, changes --zip to be a
positional argument, adds comments, updates two help strings, and adds
a general help message for users that click on the updater randomly.

Also, --delete's BooleanOptionalAction has been replaced with a simple
store_true action, as PyPlayer is now being compiled with Python 3.8 for
Windows Vista/7 compatability, and BooleanOptionalAction was added
in 3.9.