Added method of hiding PIL folder in compilation
This adds a slightly ridiculous, slightly risky method of hiding the PIL
folder by moving its contents to the root directory, importing them,
moving its contents back. This is the only way I could find to import
PIL without actually needing to keep its folder in the root directory.
Due to the potential risk, it's quite robust - a backup is used and
unlikely folder conflicts/unusual paths are accounted for.

This process only occurs the first time PIL is imported.