Added platform suffixes to "include" directory
This changes the "vlc" folder in executable/include to "vlc-windows".
The new compiling script already accounts for "-windows" and "-unix"
suffixes, autoatically deleting files/folders with non-matching
suffixes and renaming matching ones. This was originally going to
include Linux-equivalent VLC files with a "-unix" suffix, as well, but
these have been removed for the time being due to various issues,
so Linux and macOS users must still install VLC manually.