Refactored hook, improved cross-platform launching
This integrates an alternative method of determining the validity of
PID files that works on POSIX-compliant systems like Linux and macOS.
Also, the current instance of PyPlayer is now exited in a cleaner manner
by using an actual argument called "--exit", and arguments are now
parsed before logging is setup.

Additionally, running PyPlayer directly from the script will
manually import hook.py, allowing the script to be used
the same way launcher.pyw was used.