Further refactored qthelpers.py dialog modality
The way modal dialogs are handled in qthelpers.py has been refactored
further, and will now provide clearer, more accurate behavior, as well
as a fix for what appears to be a Qt bug in which setWindowModality()
is inverted (Qt.WindowModal results in a NON-modal window). All default
`modal` parameters have been reversed accordingly, and "blocking" is now
accepted as an alternative keyword in qthelpers.getDialogFromUiClass().