Fixed openPath() `explore`, added `fallback`
Before, `explore` would set `path` to its parent directory (if it
wasn't already one), which defeated the purpose of having the
`explore` parameter in the first place.

The `fallback` parameter determines whether a missing `path` will
return -1, or if its parent directory will be opened instead.
Defaults to None, and reuses `explore`'s value if so.