Added OOBE when trimming for the first time
This adds an out-of-box-experience style dialog when trimming for the
first time, which allows users to pick their trim mode. This dialog only
appears if they've never used the Trim/Fade menu before, which is
controlled by cfg.trimmodeselected. Includes a signal so that the
dialog can be called from the _save() thread.

Also, qthelpers.getDialog() now supports a `deleteOnClose` parameter.