Moved aliases to global scope, mild code-cleanup
All aliases previously defined as instance variables are now defined
globally, in the interest of optimization (with the exception of the
lambdas `is_trim_mode` and `increment_volume`). This will, however,
make the code marginally more obscure and harder to work with for
new contributors, so this change might need tweaking in the future.

Additionally, moderate code-cleanup has occurred, with imports
and startup being reorganized, comment spacing being improved,
and qthelpers.getPopup() variations being simplified (with
additional ones added).