Added more &-shortcuts and improved tab orders
This adds the &-modifier to nearly every remaining QMenu and QAction,
allowing users to hold alt and press the corresponding character as a
shortcut. These shortcuts are also dynamically added to the track and
recent file menus, and are added to the tabs in the settings dialog.
Also, the tray icon's "Show PyPlayer" QAction is now just "PyPlayer",
and the up/down buttons in the concatenation dialog now have
Ctrl+Up and Ctrl+Down shortcuts.

Additionally, the tab orders for the settings and concatenation dialogs
have been redone and improved significantly.