Added ability to copy media path to clipboard
The path is formatted by surrounding it with quotes and, if a new
setting `checkCopyEscapeBackslashes` is checked, escaping any
backslashes (making it immediately ready for Python). This option
has been added to the menubar, to the media location button (as
part of a new context menu), and to the recent file context menu.

Also, exploring media locations has been simplified and unified
in a similar manner to copying with the following functions:
`explore_media_location` and `copy_media_location`.