Added two miscellaneous settings
This adds `checkFirstFileTrayReset` and `checkRememberDropFolder`
settings, which resets `first_video_fully_loaded` when minimizing to
the system tray and updates `cfg.lastdir` when dragging-and-dropping
media. `first_video_fully_loaded` is now set to True when restarting
media in addition to it being set while opening. Also, the default
fullscreen fade-delay has been changed from 0.1 to 0.2 seconds.