Images now use QVideoPlayerLabel, improved scaling
Image files are no longer displayed through VLC, instead displayed as a
QPixmap in our custom QLabel subclass, as this allows images to open
near-instantly and allow the slider thread to stay idle. Using the
navigation keys (<-/-> and J/L by default) with an image will cycle
forward/backward. A setting for image-specific scaling has been added,
and is controlled through `QVideoPlayerLabel.updateImageScale()`.

`QVideoPlayerLabel.isCoverArt` differentiates normal images and cover
art, and is set when a file is opened based on whether or not it's a
bytes object. play() now uses a `gif` parameter to determine whether or
not to open a path as a QMovie or a QPixmap, and gifs are now loaded
as a QPixmap first to save the gif's true size (QMovie doesn't provide
access to that information) to `QVideoPlayerLabel.gifSize`. Image-
specific variants for cover art attributes/methods have been added.

Scaling/resizing has been refactored in general, now being more unified
with `QVideoPlayerLabel._resizePixmapFit(size)` and `_resizeMovieFit()`.

Gifs are now scaled based on gifSize, allowing for accurate scaling.
Changing gif-scaling while playing a gif also restarts the gif to clear
its cached frames and avoid visual artifacts. A bug with setting "Fill"
scaling on any setting causing the current image/gif/cover art to
fill the screen regardless of what setting was changed has been fixed.