Config file is now encoded, recent files use <|>
`encoding` is now a parameter accepted by several ConfigParseBetter
functions, which should fix most unicode-related errors. The current
encoding is utf-16. Also, the comma delimiter for the recent files list
has been changed to <|> (the one used for hotkeys), in order to avoid
errors on paths with commas in them.

If initial config-reading fails, the file is read again without encoding
in order to ensure backwards compatibility and compatibility with
custom config files.