Added experimental zooming support for images
This embeds our QVideoPlayerLabel onto a QScrollArea (`scrollArea`),
which is created and setup manually within setup() as Qt Designer does
not allow QScrollAreas to be initialized in the way we need.

QVideoPlayerLabel has been updated significantly, now using a custom
paintEvent not only to handle zooming but to also simplify scaling. All
scaling-related functions for images/art have been updated or removed
accordingly. Zooming is done through `QVideoPlayerLabel.setZoom()` or
`QVideoPlayerLabel.incrementZoom()`, which take a `pos` to center the
zoom around. `QVideoPlayerLabel.disableZoom()` undoes the current zoom.

Scrolling increments the current zoom value by itself divided by 6, for
a more visually consistent change between zoom levels. Zoomed images
can be dragged using similar techniques to QDraggableWindowFrame and
cropping, with a `_dragging` bool and `_draggingOffset` QPoint. Clicking
without dragging undoes the current zoom. The default zoom value is
based on the current size of the image within the player, so that the
first zoom increment is consistent with what the user already sees.
The actual way a zoom is calculated and painted depends on the scale
setting and the actual zoom value, as different techniques are used for
each situation to balance quality and performance. The default themes
now include an invisible QScrollArea style.

Also, QVideoPlayerLabel is repainted after disabling crop mode, to clear
the white background that appears behind certain transparent images.