Simplified QVideoPlayerLabel's mouse events
QVideoPlayerLabel now uses its own mouse/wheel events, and passes
these events through to underlying widgets when necessary. Additionally,
The `WA_TransparentForMouseEvents` attribute is enabled and disabled
when playing new media, based on whether or not we need the custom
mouse events or not. It is still enabled by default.