Improved resize dialog, added "Quality" QSpinBox
The resize dialog now shows the native resolution in its main label
(which has been moved to `constants.SIZE_DIALOG_DIMENSIONS_LABEL_BASE`)
and now has a QSpinBox for setting the JPEG quality of a snapshot,
which defaults to your current JPEG quality setting, and only applies
if the final snapshot is a JPEG. The quality is returned with width and
height by get_size_dialog() if `snapshot` is True, otherwise only width
and height are returned as normal.

Also, the correct version of the resize dialog now shows up when
resizing images.