Improved resize timers, modifiers, and snapping
Resize timers now wait for the user to let go of left-click before
activating, and the Ctrl/Shift modifiers are checked within timerEvent()
as opposed to in QVideoPlayer.resizeEvent(), meaning only the modifiers
being held down at the moment the resizing finishes will be considered.
This allows users to let go of modifiers or press them while actively
resizing to change the resulting behavior. Because of this more
dynamic behavior, the resize timer is now always started, regardless
of the user's snap-settings.

When QVideoPlayer.resizeEvent() detects that the window is maximized
or fullscreened, the time for that event is saved to a property called
`QVideoPlayer.last_invalid_snap_state_time`, which is used to avoid
starting a resize timer for a few tenths of a second after leaving these
states. This property is also set whenever a part of the UI's visibility
is changed, as these also trigger the resize timer.