Improved spinMinute, update_time_spins() behavior
spinMinute no longer has its maximum changed when opening media,
and update_time_spins() now no longer calls set_and_update_progress()
if the incoming frame was detected to be invalid, and instead calls
update_progress() to reset the UI to its previous state.

This fixes both the minute spin getting frozen based on how many
minutes there are past the final hour, and the "hiccup" and occurs
when attempting to set the time spins outside the boundaries of
the media.

Also, spinHour's maximum is only changed if it's going to be enabled.
This, along with the other changes, provide a very slight optimization
to open().