Added recent files settings, lowered gain to 2.0
A spinbox for controlling the maximum number of recent files,
`spinRecentFiles`, has been added. It goes from 0-1000, the
default is 25 instead of 10, and setting it to 0 disables the
tracking of recent files altogether. Additional settings for how
the number keys interact with recent files have been added, called
`checkNumKeys1SkipsActiveFiles` and `checkNumKeysUpdateRecentFiles`,
which makes pressing 1 skip the first file if it's currently playing
(defaults to False for more consistent behavior), and controls
whether or not using the number keys updates the recent files list
through the `update_recent_list` open() parameter (defaults to False
for more consistent behavior).

Recent files are now only validated (duplicates and non-existant files
removed) if the expected list is going to be less than 25 entries, and
validation stops once the required amount is reached, optimizing
situations where there are more entries than expected.

All default themes have had a QMenu section added to enable
scrollable menus for longer lists of recent files. actionClearRecent
now appears at the top of the recent files menu if there are more
than 25 files in the list, for easier access.

Also, all instances of "Note:" in tooltips have been replaced with
"NOTE:", gain has been further lowered from 3.0 to 2.0, and an
optimization has been made to marquees where they are now set
in open() using aliases and only if the media is a video.