Added `open_recent_file()`, new setting + QActions
All instances of recent-file-opening now use `open_recent_file()`, which
takes a `path` and whether or not to `update` the recent files list.

`checkRecentFilesReorderFromMenu` has been added, which controls
whether or not clicking a recent file in the menu updates its order or
not (defaults to False). QActions for opening both with and without
reordering the list have been added to the context menus for individual
recent files.