Refactored zooming, added minimum zoom settings
QVideoPlayerLabel._calculateCurrentZoom() now rounds the zoom to 4
decimal places and sets a new attribute, `QVideoPlayerLabel._baseZoom`.
_baseZoom is used to calculate the minimum permitted zoom in
QVideoPlayerLabel.setZoom(), which is now relative to the current window
size. A new setting, `spinMinimumZoomFactor`, has been added, which goes
from 0.05x to 1.00x the size of the current window (defaults to 0.25x).

Another setting, `checkForceMinimumZoom`, allows the minimum zoom
factor to change from being relative to the window to being relative to
the media (by using the zoom factor value directly instead of multiplying
it with _baseZoom) if the minimum zoom cannot be reached due to the
media being very small to begin with (defaults to True).

More QVideoPlayerLabel zooming methods now return the zoom
and key instances of zooming are now rounded to 4 decimal places.

Also, several instances of underscore case in QVideoPlayerLabel have
been replaced with camelCase, though this is not a permanent decision.