Fixed crashes, aero-snap triggering auto-resize
`last_move_time` saves the last time the window was moved. This is
used to disable the effects of the auto-resize timer if it detects that
the last move time was less than one second prior, as aero-snap
triggers moveEvent.

Also, last_window_size and last_window_pos are now set in setup(),
defaulting to None. In qtstart.after_show_setup(), if these attributes
are still None, they are given the window's size and position from
the newly loaded config. This fixes two rare instances where these
attributes are never set and crash while saving the config. As added
protection, a try-except block has been added to the lines where
these attributes are saved.