Added zoom speed settings, QGroupBox spacing
This adds the settings `spinZoomIncrement`, `spinZoomCtrlIncrement`,
and `spinZoomShiftIncrement`. These settings control the factor that the
current zoom is divided by each time the zoom is incremented, as well
as when Ctrl or Shift are held down. These default to 6, 3, and 24
(changed from 12).

Also, the spacing between group boxes in the settings dialog has been
increased significantly, to make each section more distinguishable.