Improved aero-snap, fixed audio probe/save_as typo
This improves (but does not completely fix) the interaction between
aero-snapping and auto-resizing by having the last_move_time only
be saved if there's no resize timer active or the left mouse button is
not being held down.

Additionally, an errant usage of the old "caption" save_as() parameter
has been corrected to `noun`, and calculating the size of cover art has
been updated to work appopriately with recent updates.