Fixed scaling and marquee settings during launch
This fixes several marquee settings not emitting signals after being
loaded from the config, resulting in mismatched values until the
settings were manually adjusted. This fix combined with an older
version of the fix has been moved to qtstart.after_show_setup()
to ensure the slots are using the up-to-date config values.