Don't default to a unique name for "Save" prompt
Adds `unique_default` parameters to `save_as` and `_save_as`, which
default to True. If set to False, the file browser will use the current
media's name as the default name instead of a unique version.

If the user has `checkAlwaysSaveAs` checked, this will allow them to
quickly save edits in place using "Save..." and save new files using
"Save as...".