Added `sanitize`, `ffmpeg_async`, `get_PIL_Image`
`util.sanitize()` is a simplified, optimized implementation of
`sanitize_filename.sanitize()` (which has been removed as a dependency).
`util.ffmpeg_async()` is an asynchronous version of `util.ffmpeg()` that
returns as soon as the command is called, which is used for handling the
new progress bar. `get_PIL_Image()` has been moved to `util.py` to allow
`widgets.py` to also access it.

Additionally, the two FFmpeg functions now pipe their progress using
`-progress pipe:1` to allow for their output to be captured and read for
the new progress bar.