Everything everywhere, all at once
I wish I didn't have to do it this way, but this is a year's worth of
updates and changes (most of which were done in the last 3 weeks)
compiled into a single giga-commit. Trying to keep commits organized
often takes up an absurd portion of development time, and I just don't
have the patience to sort this out into 50 different commits.

Full changelog can be found on the release page for version 0.6.0.