Reordered `_save()` operations, added comments
Eventually there should be a way to dynamically changing the order for
each save, but this change orders the operations in a way that is faster
for typical use-cases, while also preventing several potential errors.

Also, there is now a line break and an explanatory comment between each
operation, which section-marking comments before and after.