Added setting to show track titles while cycling
This adds `checkTrackCycleShowTitle` to the settings which controls
whether or not a track's title is displayed in its marquee/statusbar
while cycling (does not affect manually setting the track). Tries to
detect whether a track title is actually unique or if it's a variation
of "Track 2" or similar. If it's not unique, the generic message (i.e.
"Audio track 2 enabled") is displayed, otherwise the title is included
(i.e. "Audio track 2 'Microphone' enabled"). Defaults to `True`.

Additionally, if your track loops back to its original track (meaning
there is just one valid track to use), a special message indicating as
such is shown. This has the side-effect of preventing the player from
going black if pressing the cycle-track key while paused.

This also adds the `title` parameter to `self.set_track()`, which
allows a custom title to be passed (defaults to `None`).