Improved operation validation, added trim checks
Operations are now tracked by copying `self.operations` to a local
variable in `self.save()` and passing it along to `self._save()` instead
unqiue predetermined booleans. Instances of `operations_detected` and
its variants have been replaced, and trims/cropping are now manually
added to the dictionary.

Cropping and trimming are now validated prior to starting any work (with
the added check of making sure a trim isn't just the entire media). If
they fail validation, their keys are removed from the dictionary without
aborting the save. The save is only aborted if the dictionary is empty
after validation. This is far more scalable than my previous solution.

Additionally, the trim command is now constructed in a more readable
manner, and `self.disable_crop_mode()` now takes a `log` parameter to
control whether or not a `log_on_statusbar` call is used (to allow allow
unique messages during crop validation to not get overwritten).