Added crop info panel, improved UI spacing
The (factored) size and left/right/top/bottom values of your crop
selection are now shown and updated in real-time on the controls in the
bottom-right corner, next to the quick-checkboxes. These only appear if
there is enough space for them, and hides the quick-checkboxes early if
needed. The crop info is within a QFrame called `frameCropInfo`.

Improvements have been made to the spacing/height of the hour, minute,
and second spinboxes to have them be uniform, unchanging, and space-
efficient, and improvements have been made to window-shrinking in terms
of which elements are toggled and when.

Also, the hour spinbox no longer has its maximum value altered as
PyPlayer already natively blocks invalid values.