Added `is_pitch_sensitive_audio`/`is_bad_with_vlc`
These flags, like `is_gif` and `is_static_image`, are used to more
quickly identify when the current media requires certain behavior.

`is_bad_with_vlc` is set for any media file using the "ogg" container
format, as most (all?) codecs for it seem to play very, very poorly in
libVLC (the same behavior can be seen in VLC). NOTE: This may be
fixable by using different demuxers or something. More tests required.
This flag is currently only used to warn the user on the statusbar
that the newly played file will be unresponsive/laggy.

`is_pitch_sensitive_audio` is set for any audio mime types that do not
use the "ogg" container format, and designates that the current audio
needs to use the pitch-shift-correcting-hack. NOTE: This flag is not
actually utilized in this commit, and will instead be used in the new
high-precision progress bar update that should be one of the next few
commits.