Massively improved progress bar accuracy & timing
This adds a new thread, `high_precision_slider_accuracy_thread()`, that
is started on launch. This thread attempts to track the real-world
time that a given video/audio file has been open and unpaused, and
compares this time to the current time reported on the progress bar
when using the high-precision progress bar. This thread also handles
desyncing (rather than `fast_start_interface_thread()`, which has been
simplified accordingly). Time is tracked from the moment the current
media starts is resumes playback and is compared once per second. If
the progress bar is too far ahead, the inter-frame delay is shortened
(and vice versa) in order to ensure the progress bar lines exactly up
with reality one second in the future. If the time is desynced greater
than one second (or greater than 2 seconds from libVLC's progress),
the progress is snapped to just beyond libVLC's progress.

A new property, `ui_delay`, has been added. This is the actual
inter-frame delay that is now used by `update_slider_thread()`,
and is what gets changed by the accuracy thread. `delay` is still the
base inter-frame delay of a file, but is instead used as a reference.

Two additional accuracy properties have been added:

1. `reset_progress_offset` forces the accuracy loop to break, even if
the player is still active, and allows the comparison values (start
time, etc.) to be re-evaluated. This is used when doing something
that would throw off the comparisons, such as navigating.

2. `add_to_progress_offset` is a semi-hack - libVLC often buffers after
drastic progress jumps without actually identifying itself as buffering.
This results in the start time for the accuracy thread being very
slightly off, as libVLC didn't actually start playing immediately. This
property is added to the start time in order to better line up with
reality. This seems like it would be VERY system-dependant, but in fact,
it seems that many instances of buffering are seemingly bottlenecked
and appear to be very similar in duration on most mid/high-end systems.
Low-end systems WILL still be desynced, but it will be at most 0.1-0.3
seconds (unless it's REALLY low-end). NOTE: Due to the way this is
designed, the progress bar continues immediately but is also slowed
immediately so that it lines up one second after the buffering took
place. It will still be desynced until that second has elapsed.

Altogether, these allow the high-precision progress bar to not only
never desync (under normal circumstances), but also have near frame-
perfect accuracy for frame-seeking, often being no more than a handful
of frames off on mid/high-end systems.

---

A new method has been added, `set_and_adjust_and_update_progress`, which
serves to simplify the process of setting progress while accounting for
the accuracy thread and its new properties. It also more carefully
handles `frame_override` and the audio pitch-shift-hack. Many instances
of progress-setting have been replaced with one-line calls to this new
method, which fixes certain timing-related bugs and also fills in the
spots where I initially forgot to apply the pitch-shift-hack.

Addtionally, `frame_override` is reset to -1 in `restart()`,
a new alias "`is_high_precision_slider()`" has been added,
`QVideoSlider.mouseReleaseEvent()` has been reordered to
fix yet another timing issue, and the high-precision progress
bar's setting and tooltip have been reworded for clarity.