Added aggressive focus, revamped focus settings
This adds many new settings regarding when to focus PyPlayer. These
settings handle not only which window states should warrant focusing,
but also add additional settings for when to ignore the focus settings,
such as when there is a fullscreen application in the foreground or
audio files are being played.

Previously, in a much earlier version of PyPlayer, there was an
alternate Windows-only method of focusing that forcibly raised PyPlayer
to the top of the Z-order. This however resulted in inconsistent
behavior, often not actually focusing at all, and was removed in favor
of Qt's cross-platform APIs. However, it became very common for Windows
to block PyPlayer's attempts at focusing due to the fast-start
interface's implementation, so this commit brings back and a new, far
more reliable form of "aggressive" focusing, added as a separate
setting within the `qthelpers.showWindow` and `qthelpers.focusWindow`
functions (specified by the new `aggressive` parameter)

Aggressive focusing uses COM to send a fake Shift-key press to the
PyPlayer window before raising it with Qt, tricking Windows into
thinking `The calling process received the last input event.`
https://learn.microsoft.com/en-us/windows/win32/api/winuser/nf-winuser-setforegroundwindow

Also, `util.foreground_is_fullscreen()` has been added which returns
True if the currently focused window is fullscreen.

Additionally, `requirements.txt` now includes `pywin32`, `util.py` is
now ordered and imported in fully alphabetical order, and the
`Window behavior` section in the settings dialog has been
reorganized significantly.