Added `qthelpers.flashWindow()` and flash setting
This adds the ability to have the taskbar icon flash after opening media
without the window being raised/focused. To accomplish this, two new
functions were added: `qthelpers.flashWindow()`, which takes a variety
of parameters to accomplish different behavior (read docstring for more
information), and `qthelpers.stopFlashingWindow()`.

A new parameter has been added to `self.open`, `flash_window`, which
defaults to True, and is only considered if the window is not and will
not be focused. Right now, only opening things from the concatenation
dialog will refuse to flash the window.