Hide duplicate info from context menu, add dates
`self.add_info_actions` now skips information that is included in your
title, and now shows two dates for the media:
1. The creation date, shown both in the context menu
and the action's tooltip.
2. The last modified date, shown only in the action's tooltip.

The format these dates use is controlled by two new settings,
`lineDateFormatShort` (action) and `lineDateFormatLong` (tooltip).
If either setting is invalid, a special action is shown in the context
menu. Platform-specific codes are automatically converted.