Greatly improved concatenation dialog, new choices
The concatenation dialog has been revamped. It opens itself repeatedly
until a successful concatenation occurs or the dialog is cancelled, the
"OK" button has been split into "Save" and "Save as..." buttons (with
saving/output now following the same conventions as normal edits), a
`QToolButton` with a unique icon, `recent.png`, has been added - when
its menu is opened, the recent files list is shown, and you can click
a recent file to add it as an item; when its button is clicked, the
new `self.last_video` property (see below) is added - and keyboard
shortcuts have been added to also achieve the aforementioned behavior
(0-9 while menu is open, Ctrl+0-9 while menu is closed).

Also, new concatenation actions have been added: `actionCatBeforeLast`
and `actionCatAfterLast`. These actions immediately open the dialog with
the most recently played file and the current file as the initial items.
The most recently played file is tracked with the new `self.last_video`
property, which is updated whenever a new, non-edited file is played,
if the new `self.open` parameter `update_raw_last_file` is True. This
serves as a way to know the literal last file opened, even when the
user may be jumping around their recent files list.

Additionally, the concatenation dialog now includes a `QToolButton` with
a unique icon, `recent.png`. When clicked, `self.last_video` is added as
an item if it exists. When its menu is opened, the recent files list is
shown, and the user can click a recent file to add as an item.

Also, the concatenation submenu no longer includes an "Any" or
"None" action, instead just having a single action to immediately
open the dialog, which always includes the current (video) file if one
is being played.