Added a dedicated file-timestamp-editing dialog
This adds a new dialog for editing creation/modified/accessed times for
a given file or folder. It is relatively featureful, allowing times to
be swapped around, selectively synced with other files/folders, and
reset easily to both the original time(s) when the dialog was first
opened as well as the current actual times for the file. It also
includes a QLineEdit for inputting a timestamp format to use, which
updates in real-time and is saved to `cfg.timestampdialogformat`.
It also supports drag-and-dropping for both opening and syncing new
files/folders, or simply clicking the path at the top of the dialog
to browse for a new file/folder. And yes, despite being embedded
within PyPlayer, it fully supports (and allows you to open) any
filetype, as well as directories.

The dialog defaults to opening `self.video`. If not possible, it
automatically opens a file-browsing dialog.

This dialog can be accessed from the bottom of the "Edit" menu. To
reduce clutter, all deletion-related actions have been consolidated
into a new "Delete" submenu.

The dialog is opened through `self.show_timestamp_dialog()` (which is
quite large and may need to be split into its own file) that takes an
optional `file` parameter, and its base UI is contained within
`bin/window_timestamp.ui`.

This dialog may be uploaded to GitHub as a separate repository in the
future since it works standalone and can be easily separated from
PyPlayer (which I actually did immediately after making this - I'm
now using it in my native Windows right-click menu for files/folders).