"Fixed" flickering when navigating at end of media
This "fixes" the flickering that occurs when rapidly navigating to the
end of the current media and back again, which was caused by libVLC's
media-ending event having a minor delay, resulting in `self.restart`
being called after navigating back into the video.

This has been "fixed" with the addition of a new property,
`self.ignore_imminent_restart`, which is set to True when navigating
away from the end of a file while unpaused and not yet restarted.
When True, `self.restart` will exit early (before looping/autoplaying)
by replaying the file and restoring progress to the UI's current frame.
Additionally, the double-restarts that libVLC likes to do are now
detected and ignored in a similar way, to prevent even worse behavior.

Additionally, `self.restarted` is now reset to False while navigating.

I say "fixed" with quotes because in reality, I'm not very confident in
this change. Something about it just feels wrong and janky, and I fear
that it may be unstable, though I couldn't outright break it while
testing. It's definitely a hacky solution and may need tweaking
(or removal) in the future.