Added `self.frame_count_raw`
This fixes a semi-major oversight that has been in PyPlayer since it was
started - `self.frame_count` was the actual, raw number of frames in a
video, but the UI and all progress calculations count from 0. This means
that for a video with 9000 frames, the UI treated it from 0-9000, which
is 9001 frames. This has been the cause of a myriad of very minor but
noticable issues (involving progress calculations, editing, desync
detection, snapshots, etc.) and has finally been corrected.

The raw number of frames is now used in a new property called
`self.frame_count_raw`, while the old property now uses a UI-adjusted
value, which is the frame count minus 1 (but still always above 0 to
ensure no divide-by-zero errors).

This takes away the nice aesthetic of having `self.spinFrame` start at
0 and end at the whole frame count, but oh well. I considered keeping
the raw number in the original property and having the new property be
called `self.ui_frame_count` to keep in-line with some other property
names, but this proved too confusing as the UI-version is the one
that's needed/used for important calculations.