Hide hover-timestamp for static images, fix slider
If `QVideoSlider` is disabled, it no longer bothers painting
trim-boundaries or hover-timestamps. Also, `QVideoSlider.setMaximum()`
has been overridden and now automatically disables the slider when
`maximum` is 1 or fewer (eliminating the need to manually disable it for
static images).

Additionally, it also subtracts one from the provided `maximum` value.
This doesn't just account for the recent change to frame counts, but
also prevents the slider's handle from awkwardly jumping to the end
when going from active, normal media to a static image (or other
media with just one frame).

Additionally-additionally, `self._open_cleanup_slot` no longer manually
updates the UI to frame 0, which was only needed for static images.
This is now done at the end of `self.parse_media_file`, the same way
`self.frame_override` is set for non-images.