Added fallback system for unknown mime types
When our `filetype` library fails to detect `file`'s mime type in
`self.open`, we will now check the probe file before giving up. If
FFprobe generated a valid probe file and didn't identify the file as a
text file, we can assume it's probably a valid media file. If there's a
valid video stream, the mime type is set to "video", otherwise "audio".
Next, we compare the FFprobe `"format_name"` to a list of known problem
formats (currently just "mpegts", the format of the video where I first
encountered this issue) to determine the extension to use. If still
unknown, we check if `file` has an extension that lines up with its
presumed mime type. If still no, extension defaults to "mp4" or "mp3".

To assist with this, a new property named `self.extension_label` has
been added which specifies the actual string to be used in the titlebar.
For example, "mpeg-ts" files detected through this fallback system will
have `self.extension` set to "MP4", but `self.extension_label` set to
"MPEG-TS". For completely unknown files, the extension label will be
set to "???" to better signal to the user that PyPlayer is unsure of
what the current media actually is.