Increased usage of `stop` icon, added `x` icon
`self.stop`'s `icon` parameter now defaults to "stop", making it more
obvious at a glance that the player is stopped. I originally did
not add or use a stop-icon this way since it represents the current
state of the player, rather than what the pause button would actually
do when pressed, but this is probably worth the broken pause-button-
visual-clue-aid-detection-paradigm.

Also, the titlebar reflects this change as well, using the square
symbol when the player is in the stopped state on top of still
using it at the end of media.

Additionally, a new X-icon has been added and is being used in the
place the stop-icon was originally being used, which was when you
stopped the player, renamed/deleted/moved `self.video`, then attempted
to resume the player. This differentiates the X-icon from regular
stopping and makes it clear to the user that an error is preventing
playback.