Added backup system for the last 10 log files
Backup occurs just before logger is initialized in `qtstart.py`. Backups
are saved to a subdir named "logs" within `constants.TEMP_DIR` with the
same name as `constants.LOG_PATH`, but with the log file's last modified
time inserted before the extension. Oldest log files are deleted first
when over the 10-file limit, and age is determined by the filename, not
the actual age of each file.