Fixed progress on edits that only output "frame=1"
Some edits (such as on audio files with cover art) may output "frame=1"
repeatedly, which is not good for our progress bar. This change detects
this scenario and falls back to "out_time_ms" for progress updates.

Additionally, audio "resizing" now hints the progress bar as to what
the duration of the output will be.